Viewer refactor - move Plot and state into own file
